refactor: move common mail-preview components in components/preview
