+++ src/components/preview/attachments-block.tsx
@@ -29,31 +29,35 @@ import { legacySoapFetch } from '@zextras/carbonio-ui-soap-lib';
 import { filter, includes, map } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
-import DeleteAttachmentModal from './delete-attachment-modal';
-import { humanFileSize, isDocument, previewType } from './file-preview';
-import {
-	getAttachmentIconColors,
-	getAttachmentsDownloadLink,
-	getAttachmentsLink,
-	getLocationOrigin
-} from './utils';
-import { AppContext } from 'app-utils/app-context-initializer';
-import { getAttachmentExtension, useAttachmentIconColor } from 'helpers/attachments';
-import { openEmlStandalonePreview } from 'helpers/external-tabs';
-import { useUiUtilities } from 'hooks/use-ui-utilities';
-import { deleteAttachmentsEmailStoreAction } from 'store/emails/actions/delete-attachments-action';
+import { AppContext } from '../../app-utils/app-context-initializer';
+import { getAttachmentExtension, useAttachmentIconColor } from '../../helpers/attachments';
+import { openEmlStandalonePreview } from '../../helpers/external-tabs';
+import { useUiUtilities } from '../../hooks/use-ui-utilities';
+import { deleteAttachmentsEmailStoreAction } from '../../store/emails/actions/delete-attachments-action';
 import type {
 	AttachmentPart,
 	AttachmentType,
 	CopyToFileRequest,
 	CopyToFileResponse,
 	MailMessage
-} from 'types/index.d';
+} from '../../types';
 import {
 	ArrayOneOrMore,
 	NodeWithMetadata,
 	SelectNodesFunctionArgs
-} from 'types/integrations/carbonio-files-ui';
+} from '../../types/integrations/carbonio-files-ui';
+import DeleteAttachmentModal from '../../views/app/preview/preview/delete-attachment-modal';
+import {
+	humanFileSize,
+	isDocument,
+	previewType
+} from '../../views/app/preview/preview/file-preview';
+import {
+	getAttachmentIconColors,
+	getAttachmentsDownloadLink,
+	getAttachmentsLink,
+	getLocationOrigin
+} from '../../views/app/preview/preview/utils';
 
 /**
  * The BE currently doesn't support the preview of PDF attachments
@@ -485,7 +489,7 @@ type AttachmentsBlockProps = {
 	messageAttachments: MailMessage['attachments'];
 	isEml?: boolean;
 };
-const AttachmentsBlock = ({
+export const AttachmentsBlock = ({
 	isEml = false,
 	messageId,
 	messageSubject,
@@ -681,4 +685,3 @@ const AttachmentsBlock = ({
 		<></>
 	);
 };
-export default AttachmentsBlock;
