+++ src/components/preview/detail-panel-body-container.tsx
