+++ src/components/preview/detail-panel-body.tsx
