+++ src/components/preview/detail-panel-container.tsx
