+++ src/components/preview/detail-panel-header-content.tsx
