+++ src/components/preview/detail-panel-message-loading.tsx
