+++ src/views/app/preview/preview/tests/attachments-block.test.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 
 import { screen } from '@testing-library/react';
 
-import AttachmentsBlock from '../attachments-block';
+import AttachmentsBlock from '../../../../../components/preview/attachments-block';
 import { setupTest } from '@test-setup';
 import { useAppContext } from '@test-utils/carbonio-shell-ui/carbonio-shell-ui';
 import { previewContextMock } from '@test-utils/carbonio-ui-preview';
