+++ src/views/search/preview/message-mode/search-message-preview.tsx
@@ -8,12 +8,12 @@ import React from 'react';
 import { Container } from '@zextras/carbonio-design-system';
 import { useNavigate } from 'react-router-dom';
 
-import { MessagePanelBody } from '../../../../components/detail-panel/message-panel-body';
+import { MessagePanelBody } from '../../../../components/preview/message-panel-body';
 import { SearchPanelHeader } from '../../parts/search-panel-header';
 import { API_REQUEST_STATUS, SEARCH_ROUTE } from 'constants/index';
 import { useCompleteMessageOrFetch } from 'store/emails/hooks/hooks';
 
-export const SearchMessagePanel = ({ messageId }: { messageId: string }): React.JSX.Element => {
+export const SearchMessagePreview = ({ messageId }: { messageId: string }): React.JSX.Element => {
 	const { message, messageStatus } = useCompleteMessageOrFetch({
 		messageId
 	});
